+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- `ForgetTrackedSearcher` was not retroactively applying strategies that had a `basis`.
+
+### Changed
+- One by one verification will now only verify subclasses of the given basis.
+- Verification strategies no longer ignore parent
 
 ## [3.0.0] - 2021-06-14
 ### Added
@@ -25,6 +31,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Fixed isolated fusion bug
 - Fusion is now a one-way strategy.
 - Added length to name for `all_the_strategies` pack
+- pack for locally factorable now account for the basis.
 
 ### Fixed
 - Removed a redundant `Tiling.is_empty` check in the fusion algorithm.